requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
django-cors-headers==4.2.0
Pillow==10.1.0
python-dotenv==1.0.0
django-jazzmin==2.6.0
django-filter==23.3